requirements.txt
ultralytics
opencv-python
numpy
matplotlib
reportlab
google-generativeai
python-dotenv
